updated formTextField component
